README.MD
Farmery” is an exclusive platform for buying farming products and area specific weather report for a cropping-season. Which can be used for their betterment:
As Farmer can use this website to buy farming product , resolving their small-doubt through ChatBot and providing maintaining security and privacy subsequentially .
As Farmer need to be informed about climate, as climate in India is very unpredictable. Indian-subcontinent have four seasons of farming so we are providing weather forecast ,perhaps providing full details of the types of crops and how to cultivate them. 

Technologies Used:
1. MERN ( MongoDB, Express.js, React.js, Node.js) 
2. Stripe (Future work) - for payment purposes
3. JWT ( JSON Web Token)